+++ lib/component/user_privacy.dart
@@ -42,7 +42,7 @@ class UserPrivacyAlert extends StatelessWidget {
   Widget build(BuildContext context) => DoubleChooseWindows(
       options: FlExtended()
           .modalWindowsOptions
-          .copyWith(color: UCS.black50)
+          .copyWith(color: UCS.black50, addMaterial: true)
           .merge(options),
       decoration: BoxDecoration(
           color: UCS.white, borderRadius: BorderRadius.circular(8)),